requirements.txt
aiohttp==3.7.2
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autoflake==1.4
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
confusable-homoglyphs==3.2.0
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
Django==2.2.22
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.0.9
django-haystack==3.0
django-registration==3.1.1
django-reversion==3.0.8
django-smtp-ssl==1.0
django-sql-explorer==2.0
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
djangorestframework-simplejwt==4.6.0
ecdsa==0.14.1
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
geoip2==4.1.0
gevent==20.9.0
greenlet==0.4.17
html5lib==1.1
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.6.4
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
lockfile==0.12.2
lxml==4.6.1
MarkupSafe==1.1.1
maxminddb==2.0.3
more-itertools==8.6.0
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.1.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
pywatchman==1.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
reportlab==3.5.55
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
simplejson==3.17.2
urllib3==1.25.11
xhtml2pdf==0.2.5
xlrd==1.2.0
